+++ metric/instrument/syncint64.go
@@ -18,12 +18,18 @@ import (
 	"context"
 
 	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/metric/embedded"
 )
 
 // Int64Counter is an instrument that records increasing int64 values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Int64Counter interface {
+	embedded.Int64Counter
+
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr int64, attrs attribute.Set)
 }
@@ -64,8 +70,13 @@ type Int64CounterOption interface {
 // Int64UpDownCounter is an instrument that records increasing or decreasing
 // int64 values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Int64UpDownCounter interface {
+	embedded.Int64UpDownCounter
+
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr int64, attrs attribute.Set)
 }
@@ -107,8 +118,13 @@ type Int64UpDownCounterOption interface {
 // Int64Histogram is an instrument that records a distribution of int64
 // values.
 //
-// Warning: methods may be added to this interface in minor releases.
+// Warning: Methods may be added to this interface in minor releases. See
+// [go.opentelemetry.io/otel/metric] package documentation on API
+// implementation for information on how to set default behavior for
+// unimplemented methods.
 type Int64Histogram interface {
+	embedded.Int64Histogram
+
 	// Record adds an additional value to the distribution.
 	Record(ctx context.Context, incr int64, attrs attribute.Set)
 }